test: add model provisioner test suite covering path resolution, verification, and handoff
